third_party/WebKit/LayoutTests/external/wpt/referrer-policy/same-origin/attr-referrer/cross-origin/http-http/script-tag/cross-origin.no-redirect.http-expected.txt

This is a testharness.js-based test.
FAIL The referrer URL is omitted when a
                                 document served over http requires an http
                                 sub-resource via script-tag using the attr-referrer
                                 delivery method with no-redirect and when
                                 the target request is cross-origin. assert_equals: Reported Referrer URL is 'omitted'. expected (undefined) undefined but got (string) "http://web-platform.test:8001/referrer-policy/same-origin/attr-referrer/cross-origin/http-http/script-tag/cross-origin.no-redirect.http.html"
Harness: the test ran to completion.
